% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinGame.R
\name{calc_score}
\alias{calc_score}
\title{Calculate the payout after each round of the coin toss game}
\usage{
calc_score(
  toss_results,
  starting_capital = 100,
  head_factor = 2,
  tails_factor = 0.5,
  min_bet = 0,
  max_payout = Inf,
  betting_fraction = 1
)
}
\arguments{
\item{toss_results}{A sequence of 1s and 0s, representing heads and tails}

\item{starting_capital}{Initial amount of money}

\item{head_factor}{Factor to multiply bet by if heads}

\item{tails_factor}{Factor to multiply bet by if tails}

\item{min_bet}{Minimum bet}

\item{max_payout}{Maximum payout}

\item{betting_fraction}{Fraction of funds to bet}
}
\value{
A list of the payout after each round
}
\description{
Calculate the payout after each round of the coin toss game
}
\examples{
calc_score(0:1, 100, 2, 0.5, 1, 1000, 0.5)
}
