% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chanceGames.R
\name{calc_score_for_die}
\alias{calc_score_for_die}
\title{Calculate the payout after each round of the die game}
\usage{
calc_score_for_die(
  toss_results,
  starting_capital = 100,
  payoffs = c(2, 0.5),
  min_bet = 0,
  max_payout = Inf,
  betting_fraction = 1
)
}
\arguments{
\item{toss_results}{A sequence of integers,
representing the sides of each die roll}

\item{starting_capital}{Initial amount of money}

\item{payoffs}{a list of the payoffs, one for each side of the die}

\item{min_bet}{Minimum bet}

\item{max_payout}{Maximum payout}

\item{betting_fraction}{Fraction of funds to bet}
}
\value{
A list of the payout after each round
}
\description{
Calculate the payout after each round of the die game
}
\examples{
calc_score_for_die(1:6, 100, c(2, 0.5, 1, 1, 1, 1), 1, 1000, 0.5)
}
