% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chanceGames.R
\name{coin_toss_seq}
\alias{coin_toss_seq}
\title{Get Sequence of Coin Tosses and corresponding payouts}
\usage{
coin_toss_seq(
  rounds = 1,
  heads_prob = 0.5,
  starting_capital = 100,
  heads_factor = 2,
  tails_factor = 0.5,
  min_bet = 0,
  max_payout = Inf,
  betting_fraction = 1,
  insurance_payoffs = c(0, 1),
  use_insurance = FALSE,
  insurance_cost = 0.1
)
}
\arguments{
\item{rounds}{Number of rounds to simulate}

\item{heads_prob}{Probability of heads}

\item{starting_capital}{Initial amount of money}

\item{heads_factor}{Factor to multiply bet by if heads}

\item{tails_factor}{Factor to multiply bet by if tails}

\item{min_bet}{Minimum bet}

\item{max_payout}{Maximum payout}

\item{betting_fraction}{Fraction of funds to bet}

\item{insurance_payoffs}{a list of the payoffs for insurance,
one for each side of the coin}

\item{use_insurance}{Whether to use insurance}

\item{insurance_cost}{Cost of insurance}
}
\value{
A list with the results (list of the payout after each round),
toss results (list of coin toss results), and payout (integer)
}
\description{
Get Sequence of Coin Tosses and corresponding payouts
}
\examples{
coin_toss_seq(10, 0.5, 100, 2, 0.5, 1, 1000, 0.5)
}
