% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbs_plot.R
\name{create_side_by_side_plot}
\alias{create_side_by_side_plot}
\title{Create a side-by-side plot of time series and histogram}
\usage{
create_side_by_side_plot(df, use_log = TRUE)
}
\arguments{
\item{df}{A data frame with columns \code{round}, \code{value}, and \code{index}}

\item{use_log}{Whether to use a log scale}
}
\value{
A ggplot2 object
}
\description{
Create a side-by-side plot of time series and histogram
}
\examples{
create_side_by_side_plot(data.frame(
  round = 1:100,
  value = rnorm(100),
  index = "A"
))

}
