% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chanceGames.R
\name{die_game_seq}
\alias{die_game_seq}
\title{Generate all possible die sequences}
\usage{
die_game_seq(
  rolls,
  sides = 6,
  payoffs,
  starting_capital = 100,
  min_bet = 0,
  max_payout = Inf,
  betting_fraction = 1,
  insurance_payoffs = c(0, 1),
  use_insurance = FALSE,
  insurance_cost = 0.1
)
}
\arguments{
\item{rolls}{Number of rolls}

\item{sides}{Number of sides}

\item{payoffs}{a list of the payoffs, one for each side of the die}

\item{starting_capital}{Initial amount of money}

\item{min_bet}{Minimum bet}

\item{max_payout}{Maximum payout}

\item{betting_fraction}{Fraction of funds to bet}

\item{insurance_payoffs}{a list of the payoffs for insurance,
one for each side of the die}

\item{use_insurance}{Whether to use insurance}

\item{insurance_cost}{Cost of insurance}
}
\value{
A matrix with all possible coin toss sequences
}
\description{
Generate all possible die sequences
}
\examples{
die_game_seq(2, 6, c(2, 0.5, 1, 1, 1, 1), 100, 1, 1000, 0.5)
}
