% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lln.R
\name{draw_lln_with_func_facet}
\alias{draw_lln_with_func_facet}
\title{Get the probability that an event was generated by a normal distribution
given the alternative is a t-distribution with given degrees of freedom.}
\usage{
draw_lln_with_func_facet(n = 1000, mean_of_func, random_function, ...)
}
\arguments{
\item{n}{The prior probability that the event was
generated by a normal distribution}

\item{mean_of_func}{The Z-score of the event,
i.e., the number of standard deviations away from the mean}

\item{random_function}{The degrees of freedom for the t-distribution}

\item{...}{Additional arguments to pass to the random function}
}
\value{
a gif
}
\description{
Get the probability that an event was generated by a normal distribution
given the alternative is a t-distribution with given degrees of freedom.
}
\examples{
draw_lln_with_func_facet(400,
  mean_of_func = 2.5, rgamma, shape = 10, scale = 0.25
)
}
