% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chanceGames.R
\name{generate_die_sequences}
\alias{generate_die_sequences}
\title{Generate all possible die sequences}
\usage{
generate_die_sequences(rolls, sides = 6)
}
\arguments{
\item{rolls}{Number of rolls}

\item{sides}{Number of sides}
}
\value{
A matrix with all possible coin toss sequences
}
\description{
Generate all possible die sequences
}
\examples{
generate_coin_sequences(2)
}
