% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chanceGames.R
\name{get_coin_game_results}
\alias{get_coin_game_results}
\title{Perform the coin toss game, given a heads/tails sequence}
\usage{
get_coin_game_results(
  toss_results,
  starting_capital = 100,
  head_factor = 2,
  tails_factor = 0.5,
  min_bet = 0,
  max_payout = Inf,
  betting_fraction = 1
)
}
\arguments{
\item{toss_results}{A sequence of 1s and 0s, representing heads and tails}

\item{starting_capital}{Initial amount of money}

\item{head_factor}{Factor to multiply bet by if heads}

\item{tails_factor}{Factor to multiply bet by if tails}

\item{min_bet}{Minimum bet}

\item{max_payout}{Maximum payout}

\item{betting_fraction}{Fraction of funds to bet}
}
\value{
A list with the results (list of the payout after each round),
toss results (list of coin toss results), and payout (integer)
}
\description{
Perform the coin toss game, given a heads/tails sequence
}
\examples{
get_coin_game_results(0:1, 100, 2, 0.5, 1, 1000, 0.5)
}
