% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betting.R
\name{get_kelly_bet}
\alias{get_kelly_bet}
\title{Get the optimal fraction of your bankroll to bet based on the Kelly Criterion}
\usage{
get_kelly_bet(p = 0.5, a = 1, b = 1)
}
\arguments{
\item{p}{The probability of a positive outcome}

\item{a}{The fraction that is lost in a negative outcome.}

\item{b}{The fraction that is gained in a positive outcome.}
}
\value{
A proportion
}
\description{
Get the optimal fraction of your bankroll to bet based on the Kelly Criterion
}
\examples{
get_kelly_bet(0.8, 0.5, 2)
}
