% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstFile.R
\name{p_normal_given_event}
\alias{p_normal_given_event}
\title{Get the probability that an event was generated by a normal distribution
given the alternative is a t-distribution with given degrees of freedom.}
\usage{
p_normal_given_event(prior = 0.5, z = 10, t_df = 2)
}
\arguments{
\item{prior}{The prior probability that the event was
generated by a normal distribution}

\item{z}{The Z-score of the event,
i.e., the number of standard deviations away from the mean}

\item{t_df}{The degrees of freedom for the t-distribution}
}
\value{
A probability
}
\description{
Get the probability that an event was generated by a normal distribution
given the alternative is a t-distribution with given degrees of freedom.
}
\examples{
p_normal_given_event(0.999, 10, 2)
}
