% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chanceGames.R
\name{calc_score_for_coin_game}
\alias{calc_score_for_coin_game}
\title{Calculate the payout after each round of the coin toss game}
\usage{
calc_score_for_coin_game(
  toss_results,
  starting_capital = 100,
  heads_factor = 2,
  tails_factor = 0.5,
  min_bet = 0,
  max_payout = Inf,
  betting_fraction = 1,
  insurance_payoffs = c(0, 1),
  use_insurance = FALSE,
  insurance_cost = 0.1
)
}
\arguments{
\item{toss_results}{A sequence of 1s and 2s, representing heads and tails}

\item{starting_capital}{Initial amount of money}

\item{heads_factor}{Factor to multiply bet by if heads}

\item{tails_factor}{Factor to multiply bet by if tails}

\item{min_bet}{Minimum bet}

\item{max_payout}{Maximum payout}

\item{betting_fraction}{Fraction of funds to bet}

\item{insurance_payoffs}{a list of the payoffs for insurance,
one for each side of the coin}

\item{use_insurance}{Whether to use insurance}

\item{insurance_cost}{Cost of insurance}
}
\value{
A list of the payout after each round
}
\description{
Calculate the payout after each round of the coin toss game
}
\examples{
calc_score_for_coin_game(1:2, 100, 2, 0.5, 1, 1000, 0.5)
}
