% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinGame.R
\name{coin_toss_seq}
\alias{coin_toss_seq}
\title{Get the optimal fraction of your bankroll to bet based on the Kelly Criterion}
\usage{
coin_toss_seq(
  rounds = 1,
  heads_prob = 0.5,
  starting_capital = 100,
  head_factor = 2,
  tails_factor = 0.5,
  min_bet = 1,
  max_payout = Inf,
  betting_fraction = 1
)
}
\arguments{
\item{rounds}{Number of rounds to simulate}

\item{heads_prob}{Probability of heads}

\item{starting_capital}{Initial amount of money}

\item{head_factor}{Factor to multiply bet by if heads}

\item{tails_factor}{Factor to multiply bet by if tails}

\item{min_bet}{Minimum bet}

\item{max_payout}{Maximum payout}

\item{betting_fraction}{Fraction of funds to bet}
}
\value{
A list with the results, toss results, and payout
}
\description{
Get the optimal fraction of your bankroll to bet based on the Kelly Criterion
}
\examples{
coin_toss_seq(10, 0.5, 100, 2, 0.5, 1, 1000, 0.5)

}
