% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lln.R
\name{draw_lln_with_func_facet}
\alias{draw_lln_with_func_facet}
\title{Visualize the behavior of the Law of Large Numbers
with a given function compared to a normal distribution.}
\usage{
draw_lln_with_func_facet(
  animate = FALSE,
  n = 1000,
  theoretical_mean,
  random_function,
  ...
)
}
\arguments{
\item{animate}{Whether the returned plot should animate the draws}

\item{n}{The number of draws to be made}

\item{theoretical_mean}{The theoretical mean of the function}

\item{random_function}{The function from which samples are to be drawn}

\item{...}{Additional arguments to pass to the random function}
}
\value{
a plot or animated plot
}
\description{
Visualize the behavior of the Law of Large Numbers
with a given function compared to a normal distribution.
}
\examples{
draw_lln_with_func_facet(
  animate = FALSE, n = 400,
  theoretical_mean = 2.5, rgamma, shape = 10, scale = 0.25
)
draw_lln_with_func_facet(
  animate = FALSE, n = 400,
  theoretical_mean = 2.5, stats::rgamma, shape = 10, scale = 0.25
)
draw_lln_with_func_facet(
  animate = FALSE, n = 200,
  theoretical_mean = 2.5, stats::rcauchy, location = 2.5, scale = 2
)
draw_lln_with_func_facet(
  animate = FALSE, n = 200,
  theoretical_mean = 2.5, stats::rcauchy, location = 2.5, scale = 2
)
}
