% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinGame.R
\name{generate_coin_sequences}
\alias{generate_coin_sequences}
\title{Generate all possible coin toss sequences}
\usage{
generate_coin_sequences(num_flips)
}
\arguments{
\item{num_flips}{Number of coin flips}
}
\value{
A matrix with all possible coin toss sequences
}
\description{
Generate all possible coin toss sequences
}
\examples{
generate_coin_sequences(2)
}
